requirements/django_2_0.txt
django-autoslug-iplweb
django-formtools>=2.0
django-nine>=0.1.13
django-nonefield>=0.1
easy-thumbnails>=2.4.1
Pillow>=2.0.0
requests>=1.0.0
six>=1.9
Unidecode>=0.04.1
vishap>=0.1.5,<2.0
#https://github.com/django/django-formtools/archive/master.tar.gz
